# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bajra Books API",
    "version": "1.0.0",
    "description": "API documentation for Bajra Books - Book store platform"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Bearer token"
      }
    }
  },
  "security": [],
  "tags": [
    {
      "name": "Auth",
      "description": "Authentication and session management"
    },
    {
      "name": "Books",
      "description": "Book management"
    },
    {
      "name": "Offers",
      "description": "Offer management"
    },
    {
      "name": "Cart",
      "description": "Cart management"
    },
    {
      "name": "Orders",
      "description": "Order management"
    },
    {
      "name": "User",
      "description": "User profile and account"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstname": {
                    "type": "string"
                  },
                  "lastname": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/logged-user": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get logged in user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/book": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get all books",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "author",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "genre",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offer",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "minPrice",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "maxPrice",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/book/add": {
      "post": {
        "tags": [
          "Books"
        ],
        "summary": "Add a new book (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "author": {
                    "type": "string"
                  },
                  "isbn": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "original_price": {
                    "type": "number"
                  },
                  "discount": {
                    "type": "number"
                  },
                  "stock": {
                    "type": "number"
                  },
                  "weight": {
                    "type": "number"
                  },
                  "genre": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "cover_Img": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/book/delete/{id}": {
      "delete": {
        "tags": [
          "Books"
        ],
        "summary": "Delete a book (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/book/update/{id}": {
      "patch": {
        "tags": [
          "Books"
        ],
        "summary": "Update a book (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "author": {
                    "type": "string"
                  },
                  "isbn": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "original_price": {
                    "type": "number"
                  },
                  "discount": {
                    "type": "number"
                  },
                  "stock": {
                    "type": "number"
                  },
                  "weight": {
                    "type": "number"
                  },
                  "genre": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "cover_Img": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/book/add-fav/{id}": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Toggle book in wishlist (user)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/book/get-fav": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get wishlist (user)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/cart/add/{id}": {
      "post": {
        "tags": [
          "Cart"
        ],
        "summary": "Add book to cart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/cart": {
      "get": {
        "tags": [
          "Cart"
        ],
        "summary": "Get cart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/cart/update/{id}": {
      "patch": {
        "tags": [
          "Cart"
        ],
        "summary": "Update cart quantity",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/cart/remove/{id}": {
      "delete": {
        "tags": [
          "Cart"
        ],
        "summary": "Remove book from cart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/cart/clear": {
      "delete": {
        "tags": [
          "Cart"
        ],
        "summary": "Clear cart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/offer/create": {
      "post": {
        "tags": [
          "Offers"
        ],
        "summary": "Create an offer (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "stock-clearance",
                      "festive",
                      "custom"
                    ]
                  },
                  "discount": {
                    "type": "number"
                  },
                  "bookIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "startDate": {
                    "type": "string"
                  },
                  "endDate": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/offer": {
      "get": {
        "tags": [
          "Offers"
        ],
        "summary": "Get all offers",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/offer/{id}": {
      "get": {
        "tags": [
          "Offers"
        ],
        "summary": "Get single offer",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/offer/update/{id}": {
      "patch": {
        "tags": [
          "Offers"
        ],
        "summary": "Update offer (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "stock-clearance",
                      "festive",
                      "custom"
                    ]
                  },
                  "discount": {
                    "type": "number"
                  },
                  "addBookIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "removeBookIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "endDate": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/order/place": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Place an order",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "district": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "street": {
                    "type": "string"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "enum": [
                      "esewa",
                      "khalti"
                    ]
                  },
                  "alternativePhone": {
                    "type": "string"
                  },
                  "screenshot": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/order/get-my-orders": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get my orders",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/order/all": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get all orders (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "confirmed",
                "dispatched",
                "delivered",
                "cancelled"
              ]
            }
          },
          {
            "in": "query",
            "name": "paymentMethod",
            "schema": {
              "type": "string",
              "enum": [
                "esewa",
                "khalti"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/order/verify-payment/{id}": {
      "patch": {
        "tags": [
          "Orders"
        ],
        "summary": "Verify payment (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "paymentStatus": {
                    "type": "string",
                    "enum": [
                      "paid",
                      "rejected"
                    ]
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/order/update/{id}": {
      "patch": {
        "tags": [
          "Orders"
        ],
        "summary": "Update order status (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "dispatched",
                      "delivered",
                      "cancelled"
                    ]
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/update-profile": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Update user profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstname": {
                    "type": "string"
                  },
                  "middlename": {
                    "type": "string"
                  },
                  "lastname": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "profileImage": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}